refactor: split up `Value` impls
